refactor: OutputHandler 구현 및 적용
- Input / Output handler 유틸 기능을 enum 클래스로 만듦
- package 수정 (ui -> io.handler)
