+++ src/main/java/calculator/MainCalculator.java
@@ -1,15 +1,16 @@
 package calculator;
 
 import calculator.domain.StringCalculator;
-import calculator.ui.InputHandler;
-
-import java.util.Scanner;
+import calculator.io.handler.InputHandler;
+import calculator.io.handler.OutputHandler;
 
 public class MainCalculator {
     public static void main(String[] args) {
-        StringCalculator stringCalculator = new StringCalculator();
-
         String inputString = InputHandler.INSTANCE.readLine();
-        System.out.println(stringCalculator.calculate(inputString));
+
+        StringCalculator stringCalculator = new StringCalculator();
+        int calcResult = stringCalculator.calculate(inputString);
+
+        OutputHandler.INSTANCE.printOut(calcResult);
     }
 }
