+++ src/main/java/calculator/io/handler/OutputHandler.java
@@ -1,0 +1,9 @@
+package calculator.io.handler;
+
+public enum OutputHandler {
+    INSTANCE;
+
+    public void printOut(int calculationResult) {
+        System.out.println("Calculation result: " + calculationResult);
+    }
+}
